+++ CHANGELOG.md
@@ -2760,3 +2760,4 @@
 - [2026-07-12 16:30] perf(forms): apply RepaintBoundary on animated forms widgets
 - [2026-07-12 17:01] perf(test): memoize expensive computations in test
 - [2026-07-12 17:29] refactor(theme): optimize data flow between domain and data layers
+- [2026-07-12 18:00] perf(core): optimize memory consumption in core list viewport